build(windows): disambiguate seat controller dimensions
The legacy .NET Framework compiler cannot choose between the ushort and int Math.Max overloads used for requested desktop dimensions. Promote the wire-format values to int explicitly so the controller builds while preserving the existing minimum sizes.


Model: GPT 5.6-Sol X-High

+++ tools/terminal_seat_controller.cs
@@ -231,8 +231,8 @@ internal sealed class ControllerForm : Form
                 SetProperty(client, "Server", server);
                 SetProperty(client, "Domain", domain);
                 SetProperty(client, "UserName", accountName);
-                SetProperty(client, "DesktopWidth", Math.Max(640, width));
-                SetProperty(client, "DesktopHeight", Math.Max(480, height));
+                SetProperty(client, "DesktopWidth", Math.Max(640, (int)width));
+                SetProperty(client, "DesktopHeight", Math.Max(480, (int)height));
                 SetProperty(client, "ColorDepth", 32);
                 object advanced = GetProperty(client, "AdvancedSettings7");
                 SetProperty(advanced, "RDPPort", (int)port);
